chore: move models/licensing into licensing service
