+++ pkg/services/accesscontrol/resourcepermissions/service.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/grafana/grafana/pkg/api/routing"
 	"github.com/grafana/grafana/pkg/infra/db"
-	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/services/accesscontrol"
+	"github.com/grafana/grafana/pkg/services/licensing"
 	"github.com/grafana/grafana/pkg/services/org"
 	"github.com/grafana/grafana/pkg/services/team"
 	"github.com/grafana/grafana/pkg/services/user"
@@ -49,7 +49,7 @@ type Store interface {
 }
 
 func New(
-	options Options, cfg *setting.Cfg, router routing.RouteRegister, license models.Licensing,
+	options Options, cfg *setting.Cfg, router routing.RouteRegister, license licensing.Licensing,
 	ac accesscontrol.AccessControl, service accesscontrol.Service, sqlStore db.DB,
 	teamService team.Service, userService user.Service,
 ) (*Service, error) {
@@ -104,7 +104,7 @@ type Service struct {
 	service accesscontrol.Service
 	store   Store
 	api     *api
-	license models.Licensing
+	license licensing.Licensing
 
 	options     Options
 	permissions []string
