+++ pkg/services/featuremgmt/manager.go
@@ -6,10 +6,10 @@ import (
 	"net/http"
 	"reflect"
 
-	"github.com/grafana/grafana/pkg/infra/log"
-
 	"github.com/grafana/grafana/pkg/api/response"
+	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/models"
+	"github.com/grafana/grafana/pkg/services/licensing"
 )
 
 var (
@@ -18,7 +18,7 @@ var (
 
 type FeatureManager struct {
 	isDevMod  bool
-	licensing models.Licensing
+	licensing licensing.Licensing
 	flags     map[string]*FeatureFlag
 	enabled   map[string]bool // only the "on" values
 	config    string          // path to config file
