+++ pkg/services/featuremgmt/service_test.go
@@ -3,9 +3,10 @@ package featuremgmt
 import (
 	"testing"
 
-	"github.com/grafana/grafana/pkg/models"
-	"github.com/grafana/grafana/pkg/setting"
 	"github.com/stretchr/testify/require"
+
+	"github.com/grafana/grafana/pkg/services/licensing"
+	"github.com/grafana/grafana/pkg/setting"
 )
 
 func TestFeatureService(t *testing.T) {
@@ -47,7 +48,7 @@ func TestFeatureService(t *testing.T) {
 }
 
 var (
-	_ models.Licensing = (*stubLicenseServier)(nil)
+	_ licensing.Licensing = (*stubLicenseServier)(nil)
 )
 
 type stubLicenseServier struct {
