+++ pkg/services/licensing/models.go
@@ -1,4 +1,4 @@
-package models
+package licensing
 
 type Licensing interface {
 	// Expiry returns the unix epoch timestamp when the license expires, or 0 if no valid license is provided
